requirements.txt
blinker==1.9.0
click==8.3.2
cmdstanpy==1.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.3
fonttools==4.62.1
greenlet==3.4.0
holidays==0.94
importlib_resources==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.19.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
plotly==6.7.0
prophet==1.3.0
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
SQLAlchemy==2.0.49
stanio==0.5.1
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
Werkzeug==3.1.8
